Refactor to make use of shared/Utils.js: get/setCookie() & show/hideShakingTree()
