+++ index.html
@@ -80,7 +80,6 @@
     <script src="views/cc7/js/xlsx.full.min.js"></script>
     <script src="views/cc7/js/FileSaver.min.js"></script>
     <script type="module" src="views/cc7/js/Settings.js"></script>
-    <script type="module" src="views/cc7/js/Utils.js"></script>
     <script type="module" src="views/cc7/js/HierarchyView.js"></script>
     <script type="module" src="views/cc7/js/LanceView.js"></script>
     <script type="module" src="views/cc7/js/PeopleTable.js"></script>
