+++ views/cc7/js/HierarchyView.js
@@ -1,4 +1,4 @@
-import { Utils } from "./Utils.js";
+import { Utils } from "../../shared/Utils.js";
 
 export class HierarchyView {
     static buildView() {
