+++ views/shared/Utils.js
